Create populate_database.php
This is the php file that will load the database with data from the two APIs. I've created the first few code, declarations and function. This is the function that creates and loads the store data Please correct where necessary and update as I'm not sure if this is correct.